app/src/requirements.txt
Flask==3.1.0
google-cloud-aiplatform==1.90.0
google-cloud-texttospeech==2.26.0
moviepy==1.0.3
imageio-ffmpeg==0.6.0
google-genai==1.12.1
